requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
numpy==1.21.0
pandas==1.2.5
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
urllib3==1.26.6
